[Exceptions] Move old exceptions from types to exceptions
Put exception code in the right place.
