+++ src/aleph/types.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 from enum import Enum
 
-
-class UnknownHashError(ValueError):
-    pass
+from aleph.exceptions import UnknownHashError
 
 
 class ItemType(str, Enum):
@@ -30,8 +28,3 @@ class Protocol(str, Enum):
     """P2P Protocol"""
     IPFS = "ipfs"
     P2P = "p2p"
-
-
-class InvalidMessageError(Exception):
-    """Error raised when an invalid message is processed"""
-    pass
